+++ packages/codemod/transforms/web-ui/v0/__testfixtures__/import-paths.output.tsx
@@ -1,6 +1,5 @@
 // this is a comment at the beginning of the file
 import {
-  Background,
   useBackground,
   BackgroundProvider,
   Box,
@@ -33,8 +32,8 @@ import {
 } from '@utilitywarehouse/web-ui';
 
 import {
-  Card,
-  CardProps,
+  Background,
+  buttonClasses,
   Icon,
   IconProps,
   Hidden,
@@ -46,4 +45,4 @@ import {
 } from '@utilitywarehouse/customer-ui-material';
 
 import { useDeviceSize } from '~/hooks';
-import { Container, NavLink } from '~/components';
+import { Card, CardProps, Container, NavLink } from '~/components';