+++ packages/web-ui/src/Link/Link.stories.tsx
@@ -1,13 +1,11 @@
 import * as React from 'react';
 import { Link, LinkProps } from './Link';
 import type { Meta, StoryObj } from '@storybook/react';
-import { LegacyBackgrounds } from '../storybook-utils';
 import { Text, TextProps } from '../Text';
 import { Stack } from '../Stack';
 import { Box } from '../Box';
 import { colorsCommon } from '@utilitywarehouse/colour-system';
 import { headingVariantMapping, textVariantMapping } from '../Typography/Typography';
-import { Typography } from '../Typography';
 import { ThemeProvider } from '../ThemeProvider';
 
 const meta: Meta<typeof Link> = {
@@ -102,28 +100,6 @@ export const InlineLink: Story = {
           )
         )}
       </Stack>
-    );
-  },
-};
-
-export const LinkColor: Story = {
-  name: 'On legacy Background',
-  render: () => {
-    return (
-      <LegacyBackgrounds>
-        <Stack spacing={4}>
-          {textVariants.map(v => (
-            <Stack key={v} spacing={1}>
-              <Link href="#" variant={v as TextProps['variant']}>
-                {v} link
-              </Link>
-              <Typography component="span" variant={v as TextProps['variant']}>
-                This is a <Link href="#">link</Link>, wrapped in a {v} Typography component.
-              </Typography>
-            </Stack>
-          ))}
-        </Stack>
-      </LegacyBackgrounds>
     );
   },
 };