+++ packages/web-ui/src/utils/data-attributes.ts
@@ -8,7 +8,6 @@ export const DATA_ATTRIBUTES = {
   customColor: withDataPrefix('custom-color'),
   colorscheme: withDataPrefix('colorscheme'),
   // TODO: remove in v1
-  legacy: withDataPrefix('legacy-cwui'),
   primary: withDataPrefix('primary'),
   secondary: withDataPrefix('secondary'),
   success: withDataPrefix('success'),
@@ -17,6 +16,4 @@ export const DATA_ATTRIBUTES = {
   size: withDataPrefix('size'),
   heading: withDataPrefix('heading'),
   bgcolorBrand: withDataPrefix('bg-color-brand'),
-  // when removing this, update ToggleButton to use context
-  inverse: withDataPrefix('inverse-background'),
 };