+++ packages/codemod/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @utilitywarehouse/ds-codemod
 
+## 0.0.24
+
+### Patch Changes
+
+- [#356](https://github.com/utilitywarehouse/design-systems/pull/356) [`6a24c4e`](https://github.com/utilitywarehouse/design-systems/commit/6a24c4e551b48ab57430350eedfd715ff5db2793) Thanks [@robphoenix](https://github.com/robphoenix)! - Leave buttonClasses import from cwui
+
+## 0.0.23
+
+### Patch Changes
+
+- [#354](https://github.com/utilitywarehouse/design-systems/pull/354) [`c1a9acc`](https://github.com/utilitywarehouse/design-systems/commit/c1a9acc50cec3aad063a1d7cd6b60f72d8256ca7) Thanks [@robphoenix](https://github.com/robphoenix)! - fix
+
+## 0.0.22
+
+### Patch Changes
+
+- [#352](https://github.com/utilitywarehouse/design-systems/pull/352) [`92d0490`](https://github.com/utilitywarehouse/design-systems/commit/92d0490a5878e698d33f302c6f07a35e3a4c7441) Thanks [@robphoenix](https://github.com/robphoenix)! - Fix typography variants
+
+## 0.0.21
+
+### Patch Changes
+
+- [#350](https://github.com/utilitywarehouse/design-systems/pull/350) [`c66542c`](https://github.com/utilitywarehouse/design-systems/commit/c66542cf3dd6cfbd81575336ccfee8626972ab07) Thanks [@robphoenix](https://github.com/robphoenix)! - fix
+
+## 0.0.20
+
+### Patch Changes
+
+- [#348](https://github.com/utilitywarehouse/design-systems/pull/348) [`a5be7ca`](https://github.com/utilitywarehouse/design-systems/commit/a5be7cad6af9e21dae5eb6256095a3763a934a30) Thanks [@robphoenix](https://github.com/robphoenix)! - fix
+
+## 0.0.19
+
+### Patch Changes
+
+- [#346](https://github.com/utilitywarehouse/design-systems/pull/346) [`1cf701d`](https://github.com/utilitywarehouse/design-systems/commit/1cf701dd982cfb6af41ea6ffed7cd23883c8f595) Thanks [@robphoenix](https://github.com/robphoenix)! - Fix codemod, hopefully
+
+## 0.0.18
+
+### Patch Changes
+
+- [#343](https://github.com/utilitywarehouse/design-systems/pull/343) [`1ec17ab`](https://github.com/utilitywarehouse/design-systems/commit/1ec17ab188636502f4ab275caaa1155dad8d1cd0) Thanks [@robphoenix](https://github.com/robphoenix)! - Add Typography component codemod
+
+## 0.0.17
+
+### Patch Changes
+
+- [#338](https://github.com/utilitywarehouse/design-systems/pull/338) [`90dc1e0`](https://github.com/utilitywarehouse/design-systems/commit/90dc1e0cd822c1ff3968e8f0e3be93c33598f4ea) Thanks [@robphoenix](https://github.com/robphoenix)! - Fix backround-to-box codemod
+
+## 0.0.16
+
+### Patch Changes
+
+- [#337](https://github.com/utilitywarehouse/design-systems/pull/337) [`6af022e`](https://github.com/utilitywarehouse/design-systems/commit/6af022ea33c4dcd8300d37c3bcf9cd787af27ae7) Thanks [@robphoenix](https://github.com/robphoenix)! - Include background-to-box codemod in migration
+
 ## 0.0.15
 
 ### Patch Changes